FIX: Runtime error if input contains a standalone escape
An escape character that is not part of an ANSI escape sequence (i.e. does not match $ESCAPE_SEQUENCE_PATTERN) produces undef as the submatch because the success of the pattern matching isn't checked.
> $ nowrap
> foo^[bar
> Use of uninitialized value $esc_seq in subtraction (-) at nowrap line 147, <> line 1.
> Use of uninitialized value in addition (+) at nowrap line 148, <> line 1.
> Use of uninitialized value $append in concatenation (.) or string at nowrap line 171, <> line 1.
> Use of uninitialized value $esc_seq in subtraction (-) at nowrap line 147, <> line 1.
> Use of uninitialized value in addition (+) at nowrap line 148, <> line 1.
> Use of uninitialized value $append in concatenation (.) or string at nowrap line 171, <> line 1.

Add an additional conditional branch that handles standalone escapes, accounting for the observed behavior (in Gnome terminal) that runs of adjacent escapes plus a following non-escape character are invisible; i.e. do not contribute to the line length.